requirements.txt
typer==0.9.0
gitpython==3.1.40
#openai==1.6.1
python-dotenv==1.0.0
rich==13.7.0
httpx==0.25.2
google-generativeai==0.3.2